+++ serialinterface/portrw.py
@@ -1,27 +1,28 @@
 import threading
-
+import time
 import serial
 
 from gpsmain.interface import Interface
-
 
 def serialthread(bus, intface:Interface):
     port = serial.Serial(bus)
 
     response = []
     while threading.main_thread().is_alive():
-           next2byte=port.read(2)
-           if b'@@'== next2byte:
-               intface.addincoming(response)
+        next2byte = port.read(2)
+        if b'@@' == next2byte:
+            print('serial message recieved at: ' + str(time.time_ns()))
+            #print(response)
+            intface.addincoming(response)
+            response = []
 
-               response=[]
+        else:
+            #print(next2byte.hex())
+            response.append( next2byte.hex()[0:2])
+            response.append( next2byte.hex()[2::])
 
-
-           else:
-               #print(next2byte.hex())
-               response.append(next2byte[0:2])
-               response.append(next2byte[2::])
-
-
-           out = intface.getoutgoing()
-           port.write(out)
+        try:
+            out = intface.getoutgoing()
+            port.write(out)
+        except:
+            continue
